requirements.txt
# Core
fastapi
uvicorn[standard]
pydantic
pydantic-settings

# LLM
google-generativeai

# Data & HTTP
httpx
pandas
numpy
requests

# Vector Store & Embeddings
chromadb
sentence-transformers

# Database
motor
pymongo

# Financial Analysis
yfinance

# NLP & Text Processing
beautifulsoup4
lxml

# Utils
python-dotenv
python-dateutil
